+++ config/payment.php
@@ -82,16 +82,6 @@ return [
             'callbackUrl' => 'http://yoursite.com/path/to',
             'description' => 'payment in '.config('app.name'),
         ],
-        'pasargad' => [
-            'apiPurchaseUrl' => 'https://pep.shaparak.ir/Api/v1/Payment/GetToken',
-            'apiPaymentUrl' => 'https://pep.shaparak.ir/payment.aspx',
-            'apiVerificationUrl' => 'https://pep.shaparak.ir/Api/v1/Payment/CheckTransactionResult',
-            'terminalId' => '',
-            'merchantId' => '',
-            'certificate_path' => storage_path(env('PASARGAD_CERT_PATH', 'payment/pasargad/certificate.xml')),
-            'callbackUrl' => 'http://yoursite.com/path/to',
-            'description' => 'payment in '.config('app.name'),
-        ],
         'payir' => [
             'apiPurchaseUrl' => 'https://pay.ir/pg/send/',
             'apiPaymentUrl' => 'https://pay.ir/pg/',
@@ -130,15 +120,6 @@ return [
             'apiVerificationUrl' => 'https://sadad.shaparak.ir/VPG/api/v0/Advice/Verify',
             'key' => '',
             'merchantId' => '',
-            'terminalId' => '',
-            'callbackUrl' => 'http://yoursite.com/path/to',
-            'description' => 'payment in '.config('app.name'),
-        ],
-        'saderat' => [
-            'apiPurchaseUrl' => 'https://mabna.shaparak.ir:8080',
-            'apiPaymentUrl' => 'https://Saderat.shaparak.ir:8080/Pay/',
-            'apiVerificationUrl' => 'https://mabna.shaparak.ir:8081/V1/PeymentApi/Advice/',
-            'apiRollbackUrl' => 'https://mabna.shaparak.ir:8081/V1/PeymentApi/RollBack/',
             'terminalId' => '',
             'callbackUrl' => 'http://yoursite.com/path/to',
             'description' => 'payment in '.config('app.name'),
@@ -195,21 +176,19 @@ return [
     |
     */
     'map' => [
-        'asanpardakht' => \Shetabit\Payment\Drivers\Asanpardakht::class,
-        'behpardakht' => \Shetabit\Payment\Drivers\Behpardakht::class,
-        'idpay' => \Shetabit\Payment\Drivers\Idpay::class,
-        'irankish' => \Shetabit\Payment\Drivers\Irankish::class,
-        'nextpay' => \Shetabit\Payment\Drivers\Nextpay::class,
-        'parsian' => \Shetabit\Payment\Drivers\Parsian::class,
-        'pasargad' => \Shetabit\Payment\Drivers\Pasargad::class,
-        'payir' => \Shetabit\Payment\Drivers\Payir::class,
-        'payping' => \Shetabit\Payment\Drivers\Payping::class,
-        'paystar' => \Shetabit\Payment\Drivers\Paystar::class,
-        'poolam' => \Shetabit\Payment\Drivers\Poolam::class,
-        'sadad' => \Shetabit\Payment\Drivers\Sadad::class,
-        'saderat' => \Shetabit\Payment\Drivers\Saderat::class,
-        'saman' => \Shetabit\Payment\Drivers\Saman::class,
-        'yekpay' => \Shetabit\Payment\Drivers\Yekpay::class,
-        'zarinpal' => \Shetabit\Payment\Drivers\Zarinpal::class,
+        'asanpardakht' => \Shetabit\Payment\Drivers\Asanpardakht\Asanpardakht::class,
+        'behpardakht' => \Shetabit\Payment\Drivers\Behpardakht\Behpardakht::class,
+        'idpay' => \Shetabit\Payment\Drivers\Idpay\Idpay::class,
+        'irankish' => \Shetabit\Payment\Drivers\Irankish\Irankish::class,
+        'nextpay' => \Shetabit\Payment\Drivers\Nextpay\Nextpay::class,
+        'parsian' => \Shetabit\Payment\Drivers\Parsian\Parsian::class,
+        'payir' => \Shetabit\Payment\Drivers\Payir\Payir::class,
+        'payping' => \Shetabit\Payment\Drivers\Payping\Payping::class,
+        'paystar' => \Shetabit\Payment\Drivers\Paystar\Paystar::class,
+        'poolam' => \Shetabit\Payment\Drivers\Poolam\Poolam::class,
+        'sadad' => \Shetabit\Payment\Drivers\Sadad\Sadad::class,
+        'saman' => \Shetabit\Payment\Drivers\Saman\Saman::class,
+        'yekpay' => \Shetabit\Payment\Drivers\Yekpay\Yekpay::class,
+        'zarinpal' => \Shetabit\Payment\Drivers\Zarinpal\Zarinpal::class,
     ]
 ];
