+++ src/Events/InvoiceVerifiedEvent.php
@@ -10,23 +10,27 @@ use Illuminate\Foundation\Events\Dispatchable;
 use Illuminate\Broadcasting\InteractsWithSockets;
 use Illuminate\Contracts\Broadcasting\ShouldBroadcast;
 use Shetabit\Payment\Contracts\DriverInterface;
+use Shetabit\Payment\Contracts\ReceiptInterface;
 use Shetabit\Payment\Invoice;
 
 class InvoiceVerifiedEvent
 {
     use Dispatchable, InteractsWithSockets, SerializesModels;
 
+    public $receipt;
     public $driver;
     public $invoice;
 
     /**
      * InvoiceVerifiedEvent constructor.
      *
+     * @param ReceiptInterface $receipt
      * @param DriverInterface $driver
      * @param Invoice $invoice
      */
-    public function __construct(DriverInterface $driver, Invoice $invoice)
+    public function __construct(ReceiptInterface $receipt, DriverInterface $driver, Invoice $invoice)
     {
+        $this->receipt = $receipt;
         $this->driver = $driver;
         $this->invoice = $invoice;
     }
